backend completed

+++ backend/controllers/contactController.js
@@ -8,7 +8,7 @@ const Contact = require('../models/contactModel');
  * @access Private
  */
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -36,8 +36,20 @@ const setContact = asyncHandler(async (req, res) => {
     );
   }
 
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
   const { name, email, phone, type } = req.body;
-  const contact = await Contact.create({ name, email, phone, type });
+  const contact = await Contact.create({
+    name,
+    email,
+    phone,
+    type,
+    user: req.user.id,
+  });
   res.status(201).json(contact);
 });
 
@@ -52,6 +64,18 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(400);
     throw new Error('Contact not found');
+  }
+
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure that the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -74,6 +98,18 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(400);
     throw new Error('Contact not found');
+  }
+
+  // Check if user exists
+  if (!req.user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure that the logged in user matches the contact user
+  if (contact.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
 
   await Contact.findByIdAndRemove(req.params.id);
